+++ CHANGES.rst
@@ -21,6 +21,9 @@ Breaking changes:
 
 New features:
 
+- Added 2 imaging schema: ``plone.lead_scale_name`` and ``plone.is_lead_visible``.
+  [bsuttor]
+
 - Added ``Show Toolbar`` permission.
   [agitator]
 
@@ -594,9 +597,6 @@ New features:
 
 - Update npm package depencies.
   [thet]
-
-- Added 2 imaging schema: ``plone.lead_scale_name`` and ``plone.is_lead_visible``.
-  [bsuttor]
 
 - Supported ``remove`` keyword for configlets in controlpanel.xml.  [maurits]